Fleeing and Random Encounters
I added the SPARE menu in the battle, you can now flee from battle. I also made the player persistent so you don't need to drag it into new rooms. I also added random encounters, so the more you walk around the more likely you are to get an encounter.

HOW IT WORKS:
For the fleeing in battle there are 3 new variables: PlayerStartxPos, PlayerStartyPos and Previ_Room. When you get an encounter it sets PlayerStartxPos to your x, PlayerStartyPos to your y and Previ_Room to the current room. When you flee, it returns you to the room with the xPos and yPos that you had before going into battle.